+++ crates/plugin_resolve/src/resolver/mod.rs
@@ -30,6 +30,7 @@ enum Condition {
   Development,
   Module,
   Production,
+  Custom(String),
 }
 
 impl FromStr for Condition {
@@ -45,7 +46,7 @@ impl FromStr for Condition {
       "development" => Ok(Condition::Development),
       "production" => Ok(Condition::Production),
       "module" => Ok(Condition::Module),
-      _ => Err(format!("Invalid Condition: {}", s)),
+      c => Ok(Condition::Custom(c.to_string())),
       // _ => {}
     }
   }
@@ -622,6 +623,11 @@ impl Resolver {
         .try_file(&resolved_path_buf, context)
         .or_else(|| self.try_directory(source, &resolved_path_buf, kind, true, context))
         .unwrap_or(resolved_path);
+
+      let resolved_path = self
+        .try_browser_replace(package_json_info, &resolved_path, context)
+        .unwrap_or(resolved_path);
+
       PluginResolveHookResult {
         resolved_path,
         side_effects,