+++ examples/vite-adapter-vue/farm.config.ts
@@ -1,5 +1,9 @@
 import type { UserConfig } from '@farmfe/core';
 import vue from '@vitejs/plugin-vue';
+import AutoImport from 'unplugin-auto-import/vite'
+import Components from 'unplugin-vue-components/vite'
+import { ElementPlusResolver } from 'unplugin-vue-components/resolvers'
+import farmJsPluginSass from '@farmfe/js-plugin-sass'
 
 function defineConfig(config: UserConfig) {
   return config;
@@ -9,5 +13,32 @@ export default defineConfig({
   compilation: {
     presetEnv: false
   },
-  vitePlugins: [vue()]
+  plugins: [
+    // farmJsPluginSass()
+    '@farmfe/plugin-sass',
+    {
+      name: 'remove-css-filter-plugin',
+      priority: 0,
+      transform: {
+        filters: {
+          resolvedPaths: ['src/components/HelloWorld.vue\\?vue&(.+)&lang\\.scss']
+        },
+        executor({ content }) {
+          return {
+            content: content.replace('filter: alpha(opacity=0);', ''),
+          }
+        }
+      }
+    }
+  ],
+  vitePlugins: [
+    vue(),
+    // ...
+    AutoImport({
+      resolvers: [ElementPlusResolver({ importStyle: 'sass' })],
+    }),
+    Components({
+      resolvers: [ElementPlusResolver({ importStyle: 'sass' })],
+    }),
+  ]
 });