+++ examples/vite-adapter-vue/src/components/HelloWorld.vue
@@ -1,15 +1,89 @@
 <script setup lang="ts">
-import { ref } from 'vue'
+import { ref } from 'vue';
 
 defineProps<{ msg: string }>()
 const count = ref(0)
 console.log(count.value);
+
+interface Tree {
+  label: string
+  children?: Tree[]
+}
+
+const handleNodeClick = (data: Tree) => {
+  console.log(data)
+}
+
+const data: Tree[] = [
+  {
+    label: 'Level one 1',
+    children: [
+      {
+        label: 'Level two 1-1',
+        children: [
+          {
+            label: 'Level three 1-1-1',
+          },
+        ],
+      },
+    ],
+  },
+  {
+    label: 'Level one 2',
+    children: [
+      {
+        label: 'Level two 2-1',
+        children: [
+          {
+            label: 'Level three 2-1-1',
+          },
+        ],
+      },
+      {
+        label: 'Level two 2-2',
+        children: [
+          {
+            label: 'Level three 2-2-1',
+          },
+        ],
+      },
+    ],
+  },
+  {
+    label: 'Level one 3',
+    children: [
+      {
+        label: 'Level two 3-1',
+        children: [
+          {
+            label: 'Level three 3-1-1',
+          },
+        ],
+      },
+      {
+        label: 'Level two 3-2',
+        children: [
+          {
+            label: 'Level three 3-2-1',
+          },
+        ],
+      },
+    ],
+  },
+]
+
+const defaultProps = {
+  children: 'children',
+  label: 'label',
+}
 </script>
 
 <template>
   <h1>{{ msg }}</h1>
   <div class="card">
-    <button type="button" @click="count++">count is {{ count }}</button>
+    <el-button type="primary" @click="count++">count is {{ count }}</el-button>
+
+    <el-tree :data="data" :props="defaultProps" @node-click="handleNodeClick" />
     <p>
       Edit
       <code>components/HelloWorld.vue</code> to test HMR
@@ -24,7 +98,10 @@ console.log(count.value);
   <p class="read-the-docs">Click on the Farm and Vue logos to learn more</p>
 </template>
 
-<style scoped>
+<style scoped lang="scss">
+@forward 'element-plus/theme-chalk/src/mixins/config.scss';
+@import 'element-plus/theme-chalk/src/upload.scss';
+
 .read-the-docs {
   color: #888;
 }