+++ js-plugins/less/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @farmfe/js-plugin-less
 
+## 1.2.3
+
+### Patch Changes
+
+- a569977: only load file when query is empty and the file exists
+- Updated dependencies [a569977]
+- Updated dependencies [a569977]
+  - @farmfe/core@0.13.19
+
+## 1.2.2
+
+### Patch Changes
+
+- ddc3b40: Disable generate sourcemap for node_modules bt default
+- Updated dependencies [ddc3b40]
+  - @farmfe/core@0.13.16
+
 ## 1.2.1
 
 ### Patch Changes